21 - Enhance test database isolation and cleanup procedures

+++ tests/Functional/Repository/ExampleRepositoryTest.php
@@ -7,32 +7,55 @@ namespace CSL\Tests\Functional\Repository;
 use CSL\Entity\Example;
 use CSL\Repository\ExampleRepository;
 use CSL\Tests\Functional\KernelTestCaseBase;
+use CSL\Tests\Support\TestDatabaseGuard;
 use Doctrine\ORM\EntityManagerInterface;
 use Doctrine\ORM\Tools\SchemaTool;
 use PHPUnit\Framework\Attributes\Test;
 
 final class ExampleRepositoryTest extends KernelTestCaseBase
 {
+    private static bool $schemaCreated = false;
     private EntityManagerInterface $entityManager;
     private ExampleRepository $exampleRepository;
 
     protected function setUp(): void
     {
-        self::bootKernel();
+        $kernel = self::bootKernel();
 
         /** @var EntityManagerInterface $em */
         $em = static::getContainer()->get(EntityManagerInterface::class);
         $this->entityManager = $em;
+        TestDatabaseGuard::assertSafe($kernel->getEnvironment(), $em->getConnection());
 
         /** @var ExampleRepository $exampleRepo */
         $exampleRepo = $this->entityManager->getRepository(Example::class);
         $this->exampleRepository = $exampleRepo;
         unset($em, $exampleRepo);
 
-        $schemaTool = new SchemaTool($this->entityManager);
-        $metadata = $this->entityManager->getMetadataFactory()->getAllMetadata();
-        $schemaTool->dropSchema($metadata);
-        $schemaTool->createSchema($metadata);
+        if (!self::$schemaCreated) {
+            $schemaTool = new SchemaTool($this->entityManager);
+            $metadata = $this->entityManager->getMetadataFactory()->getAllMetadata();
+            $schemaTool->dropSchema($metadata);
+            $schemaTool->createSchema($metadata);
+            self::$schemaCreated = true;
+        }
+        $this->entityManager->getConnection()->beginTransaction();
+    }
+
+    protected function tearDown(): void
+    {
+        try {
+            if (isset($this->entityManager)) {
+                $connection = $this->entityManager->getConnection();
+                if ($connection->isTransactionActive()) {
+                    $connection->rollBack();
+                }
+                $this->entityManager->close();
+                $connection->close();
+            }
+        } finally {
+            parent::tearDown();
+        }
     }
 
     #[Test]
@@ -50,5 +73,11 @@ final class ExampleRepositoryTest extends KernelTestCaseBase
         $this->assertCount(2, $result);
         $this->assertSame('User A', $result[0]->getName());
         $this->assertSame('User B', $result[1]->getName());
+    }
+
+    #[Test]
+    public function testPreviousTestDataWasRolledBack(): void
+    {
+        $this->assertSame([], $this->exampleRepository->findAll());
     }
 }
